Fixing logic for cassandraCount
When a limit is set each Cassandra process can return rows up to that limit. When the partial results are combined any excess rows are thrown away to get the final result respecting the limit. In the instance that we want to know how many rows are returned we cannot just add up the totals from the partial results as this can be an overestimate.

In this patch I take the minimum of the limit (where defined) and the reported partial counts.

NB: I experienced the following behaviour - For small values of N (the limit) count would return N^2. For large values of N it would return a number much larger than N but smaller than N^2. This behaviour was seen with version 1.2.0 of the connector but the logic responsible for the error does not appear to have changed.